Module-4_VoiceRecognition_ConversationalEngine/Facial_Recognition_Module/Facial_Recognition_Module/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
tensorflow==2.15.0
facenet-pytorch==2.5.3
Pillow==10.1.0
numpy==1.24.3
mysql-connector-python==8.2.0
torch==2.0.1
torchvision==0.15.2
